informatics and tools optimizations

+++ src/smftools/informatics/helpers/extract_mods.py
@@ -1,6 +1,6 @@
 ## extract_mods
 
-def extract_mods(thresholds, mod_tsv_dir, split_dir, bam_suffix):
+def extract_mods(thresholds, mod_tsv_dir, split_dir, bam_suffix, skip_unclassified=True):
     """
     Takes all of the aligned, sorted, split modified BAM files and runs Nanopore Modkit Extract to load the modification data into zipped TSV files
 
@@ -9,6 +9,7 @@ def extract_mods(thresholds, mod_tsv_dir, split_dir, bam_suffix):
         mod_tsv_dir (str): A string representing the file path to the directory to hold the modkit extract outputs.
         split_dit (str): A string representing the file path to the directory containing the converted aligned_sorted_split BAM files.
         bam_suffix (str): The suffix to use for the BAM file.
+        skip_unclassified (bool): Whether to skip unclassified bam file for modkit extract command
 
     Returns:
         None
@@ -27,25 +28,31 @@ def extract_mods(thresholds, mod_tsv_dir, split_dir, bam_suffix):
         print(input_file)
         # Extract the file basename
         file_name = os.path.basename(input_file)
-        # Construct the output TSV file path
-        output_tsv_temp = os.path.join(mod_tsv_dir, file_name)
-        output_tsv = output_tsv_temp.replace(bam_suffix, "") + "_extract.tsv"
-        # Run modkit summary
-        subprocess.run(["modkit", "summary", input_file])
-        # Run modkit extract
-        subprocess.run([
-            "modkit", "extract",
-            "--filter-threshold", f'{filter_threshold}',
-            "--mod-thresholds", f"m:{m5C_threshold}",
-            "--mod-thresholds", f"a:{m6A_threshold}",
-            "--mod-thresholds", f"h:{hm5C_threshold}",
-            input_file, "null",
-            "--read-calls", output_tsv
-        ])
-        # Zip the output TSV
-        print(f'zipping {output_tsv}')
-        with zipfile.ZipFile(f"{output_tsv}.zip", 'w', zipfile.ZIP_DEFLATED) as zipf:
-            zipf.write(output_tsv, os.path.basename(output_tsv))
-        # Remove the non-zipped TSV
-        print(f'removing {output_tsv}')
-        os.remove(output_tsv)
+        if skip_unclassified and "unclassified" in file_name:
+            print("Skipping modkit extract on unclassified reads")
+        else:
+            # Construct the output TSV file path
+            output_tsv_temp = os.path.join(mod_tsv_dir, file_name)
+            output_tsv = output_tsv_temp.replace(bam_suffix, "") + "_extract.tsv"
+            if os.path.exists(f"{output_tsv}.zip"):
+                print(f"{output_tsv}.zip already exists, skipping modkit extract")
+            else:
+                # Run modkit summary
+                subprocess.run(["modkit", "summary", input_file])
+                # Run modkit extract
+                subprocess.run([
+                    "modkit", "extract",
+                    "calls", "--mapped-only"
+                    "--filter-threshold", f'{filter_threshold}',
+                    "--mod-thresholds", f"m:{m5C_threshold}",
+                    "--mod-thresholds", f"a:{m6A_threshold}",
+                    "--mod-thresholds", f"h:{hm5C_threshold}",
+                    input_file, output_tsv
+                ])
+                # Zip the output TSV
+                print(f'zipping {output_tsv}')
+                with zipfile.ZipFile(f"{output_tsv}.zip", 'w', zipfile.ZIP_DEFLATED) as zipf:
+                    zipf.write(output_tsv, os.path.basename(output_tsv))
+                # Remove the non-zipped TSV
+                print(f'removing {output_tsv}')
+                os.remove(output_tsv)
